pyproject.toml
[project]
name = "rag-general-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "alembic>=1.14.1",
    "beautifulsoup4>=4.12.3",
    "black>=24.10.0",
    "duckduckgo-search>=7.3.0",
    "faiss-cpu>=1.9.0.post1",
    "fastapi>=0.115.7",
    "langchain>=0.3.15",
    "langchain-community>=0.3.15",
    "langchain-groq>=0.2.3",
    "langchain-huggingface>=0.1.2",
    "langfuse>=2.57.13",
    "langgraph>=0.2.68",
    "lxml>=5.3.0",
    "mypy>=1.14.1",
    "passlib[bcrypt]>=1.7.4",
    "psycopg2-binary>=2.9.10",
    "pydantic[email]>=2.10.6",
    "pyjwt>=2.10.1",
    "python-dotenv>=1.0.1",
    "python-jose[cryptography]>=3.3.0",
    "python-multipart>=0.0.20",
    "ruff>=0.9.3",
    "sentence-transformers>=3.4.1",
    "sqlalchemy>=2.0.37",
    "uvicorn>=0.34.0",
]

[tool.black]
line-length = 88
skip-string-normalization = true
exclude = '''/(
  \.git
  | \.mypy_cache
  | \.venv
  | \.venv2
  | \.venv3
  | build
  | dist
  | alembic
)/'''

[tool.ruff]
line-length = 88
exclude = [
    ".git",
    ".mypy_cache",
    ".venv",
    ".venv2",
    ".venv3",
    "build",
    "dist",
    "alembic/**/*",
    "*/__init__.py"
]

[tool.ruff.lint]
select = ["E", "F", "B", "W", "I", "C", "N"]
ignore = ["E501"]
per-file-ignores = { "__init__.py" = ["F401"] }

[tool.ruff.lint.isort]
known-third-party = ["fastapi", "pydantic", "sqlalchemy"]

[tool.mypy]
strict = true
disallow_untyped_defs = false
ignore_missing_imports = true
warn_unused_ignores = true
ignore_errors = false
exclude = '.*/__init__.py'

[[tool.mypy.overrides]]
module = ["fastapi", "pydantic", "sqlalchemy"]
ignore_missing_imports = true

[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[dependency-groups]
dev = [
    "ipykernel>=6.29.5",
]

